+++ generators/aws-nfs-server/templates/aws-nfs-server/index.ts
@@ -1,0 +1,73 @@
+import {
+  getSubnet,
+  Instance,
+  VolumeAttachment
+} from "@pulumi/aws/ec2";
+import { interpolate } from "@pulumi/pulumi";
+
+import { getConfig } from "./config";
+
+export = async () => {
+  const config = await getConfig();
+
+  const options = {
+    protect: config.protect,
+    retainOnDelete: config.retainOnDelete,
+  };
+
+  const subnetId = config.subnetId;
+  const selected = getSubnet({
+    id: subnetId,
+  });
+
+  const instance = new Instance(
+    config.name,
+    {
+      ami: config.ami,
+      associatePublicIpAddress: config.associatePublicIpAddress,
+      availabilityZone: config.availabilityZone,
+      disableApiTermination: config.disableApiTermination,
+      iamInstanceProfile: config.iamInstanceProfile,
+      instanceType: config.instanceType,
+      keyName: config.keypair,
+      monitoring: config.monitoring,
+      rootBlockDevice: {
+        ...config.rootBlockDevice,
+        tags: {
+          Name: `${config.name}-root-device`,
+        },
+      },
+      subnetId: config.subnetId,
+      tags: {
+        Name: config.name,
+        ...config.tags,
+      },
+      userData: config.userData,
+      userDataReplaceOnChange: true,
+      vpcSecurityGroupIds: [config.securityGroupId as string],
+    },
+    options
+  );
+
+  if (config.volumeId) {
+    new VolumeAttachment(
+      config.name,
+      {
+        instanceId: instance.id,
+        volumeId: config.volumeId,
+        deviceName: "/dev/xvdf",
+      },
+      options
+    );
+  }
+
+  return {
+    arn: interpolate`${instance.arn}`,
+    availabilityZone: interpolate`${instance.availabilityZone}`,
+    id: interpolate`${instance.id}`,
+    name: config.name,
+    privateIp: interpolate`${instance.privateIp}`,
+    publicIp: interpolate`${instance.publicIp}`,
+    securityGroupId: config.securityGroupId,
+  };
+};